backend/requirements.txt
wheel
beautifulsoup4
Django
django-clearcache
django-environ
django-extensions
python-dotenv
psycopg2
requests
